requirements.txt
Flask==3.0.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
Pillow==10.2.0
requests==2.31.0
tensorflow==2.15.0
Ipython==8.20.0